05_Authentication && 1:N model relation

+++ posts/views.py
@@ -3,6 +3,7 @@ from django.contrib.auth.decorators import login_required
 from django.views.decorators.http import require_POST
 from .models import Post, Image
 from .forms import PostForm, ImageForm
+from django.contrib.auth.decorators import login_required
 
 
 # Create your views here.
@@ -10,12 +11,15 @@ def list(request):
     posts = get_list_or_404(Post.objects.order_by('-pk'))
     context = {'posts': posts,}
     return render(request, 'posts/list.html', context)
-    
+
+@login_required
 def create(request):
     if request.method == 'POST':
         post_form = PostForm(request.POST)
         if post_form.is_valid():
-            post = post_form.save()    # 게시글 내용 처리 끝
+            post = post_form.save(commit=False)    # 게시글 내용 처리 끝
+            post.user = request.user
+            post.save()
             for image in request.FILES.getlist('file'):
                 request.FILES['file'] = image
                 image_form = ImageForm(files=request.FILES)
@@ -34,10 +38,12 @@ def create(request):
         }
     return render(request, 'posts/form.html', context)
 
-
+@login_required
 def update(request, post_pk):
     post = get_object_or_404(Post, pk=post_pk)
-
+    if post.user != request.user:
+        return redirect('posts:list')
+        
     if request.method == 'POST':
         post_form = PostForm(request.POST, instance=post)  # 1
         if post_form.is_valid():
@@ -54,6 +60,9 @@ def update(request, post_pk):
 
 def delete(request, post_pk):
     post = get_object_or_404(Post, pk=post_pk)
+    if post.user != request.user:
+        return redirect('posts:list')
+        
     if request.method == 'POST':
         post.delete()
     return redirect('posts:list')
